affichage en cours

+++ code/v0/Header/struct.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   struct.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lazanett <lazanett@student.42.fr>          +#+  +:+       +#+        */
+/*   By: lkoletzk <lkoletzk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/12 17:31:35 by yzaoui            #+#    #+#             */
-/*   Updated: 2024/04/04 17:43:45 by lazanett         ###   ########.fr       */
+/*   Updated: 2024/04/08 16:12:36 by lkoletzk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ typedef struct s_xy
 
 typedef struct s_xyd
 {
-	double	x;
-	double	y;
+	float	x;
+	float	y;
 }	t_xyd;
 
 typedef struct s_all_texture
